refactored code to include logout button and show current logged in account
